+++ stage #3/java/module #2. REST API Basics/task02/repository/src/main/java/by/epam/esm/dao/mapper/TagRowMapper.java
@@ -14,7 +14,7 @@ public class TagRowMapper implements RowMapper<Tag> {
     @Override
     public Tag mapRow(ResultSet rs, int rowNum) throws SQLException {
         Tag tag = new Tag();
-        tag.setId(rs.getInt(TAG_ID));
+        tag.setId(rs.getLong(TAG_ID));
         tag.setName(rs.getString(TAG_NAME));
         return tag;
     }
